[VPS] Issue #49 was nearly fixed: Segmentation fault occurs irregularly in vps_test.
When C and python are integrated, there seems to be a bug between some
packages. For example, scipy's loadmat / savemat caused a segmentation
fault in python + 'C ++' environment.

Fixed 1 : Instead of saving the db image list as a mat file, modify them into
 dictionary variables.

Todo : In addition, in the case of requests.get(), when it is called within
 complex python code, it causes a segmentation fault, but it has not yet
 been confirmed which package is in conflict with.